refactor roles and access libs: split roles into separate files; add isSet() and isMap() functions to utils/extras; refactor and describe rules.js.
